Reorganize canonicalize_ireq and PIP_VERSION into utils
canonicalize_ireq now returns a copy ireq instead of modifying in place
